reference_files/requirements.txt
# Core Web Framework
fastapi==0.112.2
uvicorn==0.30.6
python-multipart==0.0.9  # For FastAPI file uploads
python-dotenv==1.0.1     # For environment variables

# Database and Supabase
supabase==2.10.0         # Main Supabase client (includes storage3, postgrest, etc.)
psycopg2-binary==2.9.9   # PostgreSQL adapter
pgvector==0.2.5          # For vector operations

sendgrid==6.11.0

# AI/LangChain Stack
langchain==0.3.11
langgraph[all]==0.2.35
langchain-anthropic==0.3.0
langchain-core==0.3.24
langchain-openai==0.2.6
anthropic==0.39.0
openai==1.54.3
langchain-community==0.3.5
langchain-postgres==0.0.12
langchain-text-splitters==0.3.0
langgraph-checkpoint==2.0.1
langgraph-checkpoint-postgres==2.0.1
langsmith==0.1.129

# Zoom Integration
httpx==0.27.2           # Modern HTTP client (used by multiple components)
httpx-sse==0.4.0        # Server-Sent Events support

# Slack Integration
slack_bolt==1.20.1
slack_sdk==3.31.0

# Redis
redis>=4.2.0

# Email validation
pydantic[email]

# Utils
pydantic==2.8.2         # Data validation
tenacity==8.5.0         # Retry logic
PyYAML==6.0.2           # YAML parsing
gunicorn==23.0.0        # Production server
backoff==2.2.1

# Operational
APScheduler==3.10.4     # For scheduled tasks
PyJWT==2.10.1

resend
